refactor: split up PlantingElement across files, rename to Planting
